+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+* Command-line interface for downloading and listing lexicons ([#47])
+
 ### Fixed
 
 * Cast `pathlib.Path` to `str` for `sqlite3.connect()` ([#58])
@@ -19,4 +23,5 @@ abandoned, but this is an entirely new codebase.
 [v0.1.0]: ../../releases/tag/v0.1.0
 [unreleased]: ../../tree/main
 
+[#47]: https://github.com/goodmami/wn/issues/47
 [#58]: https://github.com/goodmami/wn/issues/58